refactor: Move Context object to new class (#929)
All logic associated with the creation of the context object resides
inside the Indexer class, which bloats the file and complicates unit
testing. In addition, it prevents access to the context object itself so
that these calls can be made through the Indexer class itself.

This PR refactors this logic into its own class, allowing access to
context methods through Indexer. The object is still recreated each time
it is accessed as the object bakes in some state information (block
height and log entry array) which need to be updated on subsequent
accesses.

I've also reorganized the folders so that we have fewer folders under
src as it is getting cluttered.