feat: Idempotent Provisioning (#1004)
Provisioning is occasionally finnicky, failing on different parts of the
process. Having retryable provisioning allows us to heal indexer state
through retries rather than manually actioning this. To this, the
provisioning step needs to have access to accurate status of
provisioning in order to make the correct decisions on which
non-idempotent actions to take.

This PR introduces a ProvisioningState class which represents the status
of provisioning, and can be used to make decisions on what parts of
provisioning need to be completed.