refactor: Move synchronisation logic to `LifecycleManger` (#1005)
This PR moves the synchronisation logic within
`ExecutorsHandler`/`BlockStreamsHandler` in to `LifecycleManager`. The
motive behind this is to provide greater awareness/control over the
individual sync tasks (managing unhealthy, reconfiguring, etc.) to
`LifecycleManger`.

I plan to build on top of this in my next PR, and have
`LifecycleManager` count how many times an unhealthy stream/executor was
restarted, and move the Indexer to `Suspended` if it goes above a
configured threshold. This refactor makes "counting" actually possible.

Wanted to land the refactor PR first to make review easier.